Store provider records in separate LevelDB datastore (#62)
This PR configures the Docker image to store the provider records in a
local LevelDB datastore.

The DHT provider records (keys starting with the `/providers` prefix)
are currently overwhelming the database with millions of records in the
`datastore` table. Queries running on the table take a lot of time and
block other goroutines, resulting in OOM kills over time.

We don't really need to persists these records as they can be rebuilt
from the network if the IPFS node restarts. Thus we can store this
records in a separate local LevelDB datastore.